+++ tests/common/fixtures/conn_graph_facts.py
