+++ tests/common/platform/transceiver_utils.py
